better many things look at the commit for more info
